- Refactored Mode and subclasses of Mode - Reorganized file locations
